refactor(youtube/general-ads-patch): squash extension classes (#1006)
